backend/requirements.txt
fastapi>=0.135.0
uvicorn[standard]>=0.41.0
pillow>=12.0.0
aiosqlite>=0.22.0
python-multipart>=0.0.22
bcrypt>=4.3.0
PyJWT>=2.10.1
